chore: upgrade gqlgen and DCS-gRPC bindings, split resolver helpers
The previous commit worked around gqlgen v0.17.47 failing silently on modern
Go toolchains by telling developers to set GOTOOLCHAIN. That is not a real
fix: a `toolchain` directive in go.mod cannot pin downward, so the local Go
still wins, and remembering an env var per shell is a trap.

Upgrade gqlgen to v0.17.94 instead, which generates cleanly on current Go, and
take DCS-gRPC go-bindings to v0.8.1 to match the 0.8.1 server release.

gqlgen owns and rewrites graph/resolver.go, which is where the aggregation
helpers happened to live, so move them to graph/breakdown.go. Rename the
ShotsByPlayer argument to match the schema, since gqlgen regenerates that
signature and the old `pID` name collided with the parsed local.

go.mod now requires Go 1.25, so move the builder image to golang:1.25-alpine
and have CI read the version from go.mod rather than pinning it separately.
